+++ .github/workflows/kona_compile.yaml
@@ -14,4 +14,4 @@ jobs:
       uses: actions/checkout@v3
     - name: Compile the client program
       run: cargo build --profile release-client-lto
-      working-directory: client-programs/zkvm-client
+      working-directory: zkvm-client
