+++ .gitignore
@@ -12,7 +12,9 @@ data
 
 # .env
 .env
-.env.server
+
+# elf
+**elf
 
 # txt
 **.txt
