--- client-programs/fault-proof/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "fault-proof"
-version = "0.1.0"
-edition.workspace = true
-authors.workspace = true
-license.workspace = true
-repository.workspace = true
-homepage.workspace = true
-
-[dependencies]
-
-# workspace (general)
-cfg-if.workspace = true
-
-# workspace (ethereum)
-alloy-consensus.workspace = true
-
-# workspace (kona)
-kona-primitives.workspace = true
-kona-executor.workspace = true
-kona-client.workspace = true
-
-# workspace (local)
-client-utils.workspace = true
-
-# sp1
-sp1-zkvm = { workspace = true }
-
-[dev-dependencies]
