--- contracts/foundry.toml
@@ -1,10 +1,0 @@
-[profile.default]
-src = "src"
-out = "out"
-libs = ["lib"]
-ffi = true
-fs_permissions = [
-    { access = "read", path = "zkconfig.json" }
-]
-
-# See more config options https://github.com/foundry-rs/foundry/blob/master/crates/config/README.md#all-options
