--- contracts/test/ZKL2OutputOracle.t.sol
@@ -1,12 +1,0 @@
-// SPDX-License-Identifier: MIT
-pragma solidity ^0.8.0;
-
-import { Test, console } from "forge-std/Test.sol";
-
-contract ZKL2OutputOracleTest is Test {
-    // get sample proof and use it to...
-    // 1) prove a proposal works
-    // 2) prove it fails with wrong block num
-    // 3) failed without checkpoint, succeeds with
-    // 4) on fork, withdrawal from Portal
-}
