+++ crates/client-utils/Cargo.toml
@@ -30,4 +30,10 @@ serde.workspace = true
 rkyv.workspace = true
 sha2.workspace = true
 
+# sp1
+sp1-zkvm.workspace = true
+
 log.workspace = true
+
+# # kzg
+# kzg-rs.workspace = true
