--- prover-server/README.md
@@ -1,10 +1,0 @@
-# Prover Server
-
-A proving server that can request batch and aggregate proofs from a server. Also exposes a route
-to get the status of a proof request.
-
-## Usage
-
-```bash
-cargo run --bin server --release
-```
