+++ zkvm-host/Cargo.toml
@@ -12,6 +12,10 @@ name = "single"
 path = "bin/single.rs"
 
 [[bin]]
+name = "fetch_and_save_proof"
+path = "bin/fetch_and_save_proof.rs"
+
+[[bin]]
 name = "multi"
 path = "bin/multi.rs"
 
@@ -19,12 +23,18 @@ path = "bin/multi.rs"
 name = "agg"
 path = "bin/agg.rs"
 
+[[bin]]
+name = "server"
+path = "bin/server.rs"
+
 [dependencies]
 
 # workspace
 tokio = { workspace = true }
 clap = { workspace = true }
 cargo_metadata = { workspace = true }
+alloy-consensus = { workspace = true }
+alloy-primitives = { workspace = true }
 
 # kona
 kona-host = { workspace = true }
@@ -37,7 +47,16 @@ sp1-sdk = { workspace = true }
 anyhow.workspace = true
 dotenv.workspace = true
 num-format.workspace = true
+revm.workspace = true
 client-utils.workspace = true
+serde = { workspace = true }
+
+# server
+axum = "0.7.4"
+bincode.workspace = true
+futures.workspace = true
+serde_cbor.workspace = true
+log.workspace = true
 
 [build-dependencies]
 sp1-helper = { workspace = true }
