+++ zkvm-host/bin/multi.rs
@@ -5,15 +5,13 @@ use clap::Parser;
 use client_utils::precompiles::PRECOMPILE_HOOK_FD;
 use host_utils::{
     fetcher::{ChainMode, SP1KonaDataFetcher},
-    get_proof_stdin,
-    precompile::precompile_hook,
-    ProgramType,
+    get_proof_stdin, ProgramType,
 };
 use kona_host::start_server_and_native_client;
 use sp1_sdk::{utils, ExecutionReport, ProverClient};
-use zkvm_host::stats::{BnStats, ExecutionStats};
+use zkvm_host::{precompile_hook, BnStats, ExecutionStats};
 
-pub const MULTI_BLOCK_ELF: &[u8] = include_bytes!("../../elf/range-elf");
+pub const MULTI_BLOCK_ELF: &[u8] = include_bytes!("../../elf/validity-client-elf");
 
 #[derive(Parser, Debug)]
 #[command(author, version, about, long_about = None)]
